Split pre_tasks.yml to portion for Linux and WSL
Linux part will do usual stuff either on bare metal _or_ in WSL distro
while WSL will detect (on Linux side) if it is WSL, which version and
attempts to setup WinRM (if not already set). Then adds host to
inventory dynamically so that Windows related playbooks can be executed.