Rework Neovim packs
Packs are not loaded automatically now. Instead they are put in opt/ and
loaded on demand by pack.vim.

In conf/ snippets can be pasted that run either _pre (if suffixed as
such) loading the pack or post loading it. This is convenient for both
Vimscript plugins that usually require globals set before loading, as
well as Lua plugins that are configured by calling .setup() function.

This also adds nvim-colorizer pack to Ansible.