Check if WSL interop works
Due to bug in WSL (presumably) interop currently does not works which
means WSL is unable to run Windows binaries. is_wsl relies on that
feature heavily so adding this check that disables is_wsl completely if
interop does not work to at least carry on with Linux side stuff.